Fix ES818 corrective factors - the older version still uses shorts for the component sum

+++ server/src/main/java/org/elasticsearch/index/codec/vectors/es818/ES818BinaryQuantizedVectorsReader.java
@@ -469,7 +469,7 @@ public class ES818BinaryQuantizedVectorsReader extends FlatVectorsReader impleme
             binarizedQueryData.writeInt(Float.floatToIntBits(r.lowerInterval()));
             binarizedQueryData.writeInt(Float.floatToIntBits(r.upperInterval()));
             binarizedQueryData.writeInt(Float.floatToIntBits(r.additionalCorrection()));
-            binarizedQueryData.writeInt(r.quantizedComponentSum());
+            binarizedQueryData.writeShort((short) r.quantizedComponentSum());
         }
         return docsWithField;
     }
